pc_app/requirements.txt
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
pyserial==3.5
Pillow==10.3.0
cairosvg==2.7.1
